Update lab to test authentication with has_secure_password
I tried to leave the lab relatively untouched except for
updating authentication and some tests around auth. This means,
for example, that the stuff in LoginHelper hasn't been changed
(even though it probably should be at some point -- but it's
likely to break a lot of things in the meantime).